# backend-services/go/src/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Simple Service API sample",
        "title": "Simple Service API",
        "termsOfService": "http://0.0.0.0:3010/swagger/doc.json",
        "contact": {},
        "version": "1.0"
    },
    "host": "0.0.0.0:3010",
    "basePath": "/api/v1/sas",
    "paths": {
        "/auth": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Hello from a private endpoint! You need to be authenticated to see this.",
                "produces": [
                    "application/json"
                ],
                "summary": "Authenticated Endpoint",
                "operationId": "auth",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Please provide a valid API key.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/public": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This route is always accessible.",
                "produces": [
                    "application/json"
                ],
                "summary": "Public Endpoint",
                "operationId": "public",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rbac": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Hello from a private RBAC restricted endpoint! You need the read:messages scope to see this.",
                "produces": [
                    "application/json"
                ],
                "summary": "RBAC restricted Endpoint",
                "operationId": "rbac restricted",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Please provide a valid API key.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Permission denied.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}